Better diagnostics for bounce notifications
This is a fair bit of refactor/cleanup for Ophan Housekeeper - it
was precipated by my trying to debug the source of a raised level
of bounce messages, spending quite some time before I realised it
wasn't Ophan! It was Data Technology's instance of Airflow...

The new logging messages should be graphable in the Mighty ELK,
and from the mail 'source' (the email address of the sender) it
should be easy to determine which project is responsible.

I've done a few other things:

* Switched to play-json case-class-based parsing, as we had an
  increasing number of fields we needed to read to give good diagnostics.
* Slightly more sophisticated use of Scanamo to give some more concise
  code. Also, modelled fewer fields from the `ophan-alerts` table as
  we don't need the others.
* Additional tests to ensure we correctly parse the different kind of
  messages we're going to get
* After a lot of struggles got JSON-based logging going
  https://github.com/guardian/cloudwatch-logs-management#scala-lambdas
* Increased the memory allocation of the Lambda, trying to address
  the Metaspace java.lang.OutOfMemoryError exceptions we've been seeing
  https://github.com/guardian/ophan-housekeeper/issues/4

See also https://github.com/guardian/ophan-data-lake/pull/4486
for the action to update the email addresses Airflow uses.