Fix Lambda PolicyDocument so Housekeeper can access DynamoDB
The old cloudformation was incorrectly formatted - we didn't get any
visible warning of this when we applied it, instead it just gave us
a policy that only had the SNS permission, not the DynamoDB permission
or the (superfluous) `lambda:InvokeFunction` - looks like it just
found the last `Statement` and used that:

https://user-images.githubusercontent.com/52038/74730343-d815a000-523d-11ea-825b-cd64f0223605.png

This update fixes the formatting, removes the unnecessary
`lambda:InvokeFunction` (which grants the right to invoke any other
lambda in the account), and also narrows the permitted DynamoDB actions
to just what the Lambda needs - the ability to Query & Delete.

With a test deploy of this Cloudformation update in place, Ophan
Housekeeper was finally able delete the Ophan Alerts which had
been sent to a permanently bouncing address:

https://logs.gutools.co.uk/s/ophan/app/kibana#/doc/581f7b00-c4df-11e9-bc08-9d1af4d9c1d5/logstash-ophan-2020.02.18?id=kUgmWHABBMxQt1Il90L6&_g=()

> ...successfully deleted 24 alerts for 'xxx.yyy@guardian.co.uk'}

Housekeeper also logged that it was going to send a message to the SNS
topic:

https://logs.gutools.co.uk/s/ophan/app/kibana#/doc/581f7b00-c4df-11e9-bc08-9d1af4d9c1d5/logstash-ophan-2020.02.18?id=FUgmWHABBMxQt1Il90L4

However, despite being subscribed to the SNS Topic, I didn't recieve
any notification email, so I suspect there is still some kind of bug
going on there - possibly the lambda runtime ending before it has a
chance to complete the action?